perplexity/requirements.txt
httpx
beautifulsoup4
python-dotenv
openpyxl
playwright
boto3

